requirements.txt
customtkinter
pygame
yt_dlp
youtube_search
pillow
typing